+++ stac/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+- Removed unnecessary qualifications ([#251](https://github.com/stac-utils/stac-rs/pull/251))
+
 ## [0.7.0] - 2024-04-29
 
 ### Added