+++ stac/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.7.2] - 2024-07-24
+
 ### Fixed
 
 - `Bounds::to_geometry` ([#279](https://github.com/stac-utils/stac-rs/pull/279))
@@ -304,7 +306,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 Initial release.
 
-[Unreleased]: https://github.com/stac-utils/stac-rs/compare/stac-v0.7.1...main
+[Unreleased]: https://github.com/stac-utils/stac-rs/compare/stac-v0.7.2...main
+[0.7.2]: https://github.com/stac-utils/stac-rs/compare/stac-v0.7.1...stac-v0.7.2
 [0.7.1]: https://github.com/stac-utils/stac-rs/compare/stac-v0.7.0...stac-v0.7.1
 [0.7.0]: https://github.com/stac-utils/stac-rs/compare/stac-v0.6.0...stac-v0.7.0
 [0.6.0]: https://github.com/stac-utils/stac-rs/compare/stac-v0.5.3...stac-v0.6.0